+++ packages/app-desktop/src/electron/channels.ts
@@ -30,4 +30,5 @@ export enum ChannelNames {
   RESOLVE_EMBED = "resolve-embed",
   UPLOAD_EMBED = "upload-embed",
   UPLOAD_EMBED_WITH_BUFFER = "upload-embed-with-buffer",
+  GET_ACCENT_COLOR = "get-accent-color"
 }